Removed redundant swmt in favor of generalized wmt
- Removed `swmt.py` and the `SurfaceWaterMassTransformation` class
within it, with the idea that all of its functionality is already
included within the core of `wmt.py` (and now `wm.py`).

- Moved the core (parent) `WaterMass` class to `wm.py`. This was
actually done in the last two commits, but here we include the
associated updates to the exporting in `xwmt.__init__.py`. The
keyword argument `t_name='temperature'` is now passed to `WaterMass`
in the analytical convergence tests, which had broken.

- Generalized core `WaterMassTransformations.datadict` method to
now identify surface fluxes as tendencies that do not have any
vertical coordinates (taken from `self.grid.axes['Z'].coords`)
in their dimensions. Addresses

- The vertical coordinate used in both calculations for pressure
(in the `wm.WaterMass.get_density` method) and for the 2D->3D
expansion that is currently used to compute WMT from surface fluxes
(see `compute.expand_surface_to_3d`) has been renamed to `z` (with
levels `z_l` and interfaces `z_i`) to emphasize that these only
support depth coordinates as currently implemented. For WMT calculations
in arbitrary vertical coordinates, these will need to be fundamentally
changed. (See related discussion at https://github.com/NOAA-GFDL/xwmt/issues/3).

- Added `/conventions/MOM6.yaml` file that can be read in as a
python dictionary that is formatted as required by the
`budgets_dict` argument of `WaterMassTransformations`, and
includes the default names of the core heat, salt, and mass
tendencies required for full WMT calculations in temperature,
salinity, and density space.

Issues addressed by this commit (or previous ones building up to)
- Makes many of the re-labellings requested in https://github.com/NOAA-GFDL/xwmt/issues/28

- Partially addresses https://github.com/NOAA-GFDL/xwmt/issues/13, although the information
is still stored as a dictionary. It should not be fairly straight forward to encode this information
in attributes of the data arrays in `self.ds`.

- Completely addresses https://github.com/NOAA-GFDL/xwmt/issues/7 and https://github.com/NOAA-GFDL/xwmt/issues/8,
I think, except for the expansion of 2D surface fluxes (see above).
There are now no redundant function or method definitions, as we
have removed `swmt.py` completely and made use of class inheritance.