Streamlined WMT calculations by reconciling tracer and mass flux computations
The new `budgets_dict` comprehensive budget format now includes explicit instructions for
how to convert each term into the right units, and how it relates to other diagnostics or
terms in the tracer budget. This means that 2D and 3D fields can now be treated essentially
the same way in the code, such that it no longer requires passing awkward attribute dictionaries
or a bunch of conditions to handle different cases.

`conventions/MOM6.yaml` was removed because the improved version now lives in `xbudget`, which is
a dependency of `xwmt`.

Notebook will be replaced in the following commit.