Support calculations with 2D surface fields only
Uses trick from https://github.com/NOAA-GFDL/xwmt/blob/07b2e6fc95fd5bb715b461656509a8497c21b87a/xwmt/swmt.py
Essentially, if Z coordinate is missing from provided dataset, then it creates a dummy Z-coordinate for a
thin upper layer (1 m thickness, not that it matters). Assuming there is no corresponding flux out of this layer
(which is slightly incorrect for temperature), the surface flux is equivalent to the vertically-integrated flux
convergence. This flux convergence is then binned in surface-lambda space (which is almost expanded to 3D just
so that the same functions can be used as for the full 3D calculations).

Part 4 of the example notebook shows that with daily output, the errors due to offline regridding and ignoring
penetrating surface fluxes (like shortwave) are relatively small.