Reverse sign of WMTs and fix support for 2D surface fluxes
We have changed the sign of WMTs such that they are positive when they
increase the mass of water with tracer values less than the threshold.
This is a breaking change, but earlier results can be recovered by
simply flipping the sign of the results. The package now agrees
with the sign convention in the original Walin 1982 paper.

Some other minor changes:
-Bumped the minor version number because of quasi-breaking change in
sign of transformation rates
-Add condition so that method still works for 2D (surface flux) diagnostics
-Rename LHS and RHS terms from "material_derivative" to "transformation"

Update all notebooks accordingly