Specify autoparse_metadata kwargs in xgcm.Grid
There was a breaking change in the default behavior of xgcm's API,
I think after this  PR https://github.com/xgcm/xgcm/pull/559. The
new default behavior is to try to infer axis info from metadata.

This retains the old behavior by specifying the `autoparse_metadata = False`
kwargs so that `xgcm.Grid` knows we are specifying the metadata ourselves